requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
antlr4-python3-runtime==4.9.3
apturl==0.5.2
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
audioread==3.0.1
av==12.0.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
click==8.0.3
colorama==0.4.4
coloredlogs==15.0.1
colorlog==6.8.2
command-not-found==0.3
contourpy==1.2.1
cryptography==3.4.8
ctranslate2==4.2.1
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
decorator==5.1.1
deepl==1.18.0
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
docopt==0.6.2
einops==0.8.0
faster-whisper==1.0.2
ffmpeg==1.4
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.5.0
greenlet==3.0.3
httplib2==0.20.2
huggingface-hub==0.23.0
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.4
joblib==1.4.2
julius==0.2.7
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.2.4
lightning-utilities==0.11.2
llvmlite==0.42.0
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
netifaces==0.11.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
omegaconf==2.3.0
onnxruntime==1.17.3
optuna==3.6.1
packaging==24.0
pandas==2.2.2
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.2
pooch==1.8.1
primePy==1.3
protobuf==5.26.1
ptyprocess==0.7.0
pyannote.audio==3.2.0
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
pytorch-lightning==2.2.4
pytorch-metric-learning==2.5.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
regex==2024.5.15
reportlab==3.6.8
requests==2.25.1
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
semver==3.0.2
sentence-transformers==2.7.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
speechbrain==1.0.0
SQLAlchemy==2.0.30
sympy==1.12
systemd-python==234
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.0
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.3.0
torchmetrics==1.4.0.post0
tqdm==4.66.4
transformers==4.40.2
triton==2.3.0
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
websockets==12.0
xdg==5
xkit==0.0.0
yarl==1.9.4
zipp==1.0.0
